Prevent health check failures from restarting buildkit
